fix: crash on Android 14 QPR3 /15 Use cursorView instance in user app service instead of remote server process

+++ app/src/main/java/xtr/keymapper/server/IInputInterface.java
@@ -16,8 +16,8 @@ public interface IInputInterface {
     MouseEventHandler getMouseEventHandler();
     KeymapProfile getKeymapProfile();
     IRemoteServiceCallback getCallback();
-    void moveCursorX(float x);
-    void moveCursorY(float y);
+    void moveCursorX(int x);
+    void moveCursorY(int y);
     void hideCursor();
     void showCursor();
 }
